+++ wand/image.py
@@ -368,6 +368,11 @@ class Image(Resource):
 
     """
 
+    #: (:class:`Metadata`) The metadata mapping of the image.  Read only.
+    #:
+    #: .. versionadded:: 0.3.0
+    metadata = None
+
     c_is_resource = library.IsMagickWand
     c_destroy_resource = library.DestroyMagickWand
     c_get_exception = library.MagickGetException
@@ -460,6 +465,7 @@ class Image(Resource):
                     read = True
                 if not read:
                     raise TypeError('invalid argument(s)')
+            self.metadata = Metadata(self)
             self.sequence = Sequence(self)
         self.raise_exception()
 
@@ -1423,6 +1429,51 @@ class Iterator(Resource, collections.Iterator):
         return type(self)(iterator=self)
 
 
+class Metadata(collections.Mapping):
+    """Class that implements dict-like read-only access to image metadata
+    like EXIF or IPTC headers.
+
+    :param image: an image instance
+    :type image: :class:`Image`
+
+    .. versionadded:: 0.3.0
+
+    """
+
+    def __init__(self, image):
+        if not isinstance(image, Image):
+            raise TypeError('expected a wand.image.Image instance, '
+                            'not ' + repr(image))
+        self.image = image
+
+    def __getitem__(self, k):
+        """
+        :param k: Metadata header name string.
+        :type k: :class:`basestring`
+        :returns: a header value string
+        :rtype: :class:`str`
+        """
+        if not isinstance(k, basestring):
+            raise TypeError('k must be a string, not ' + repr(format))
+        v = library.MagickGetImageProperty(self.image.wand, k)
+        if v is None:
+            raise KeyError
+        return v
+
+    def __iter__(self):
+        num = ctypes.c_size_t()
+        props_p = library.MagickGetImageProperties(self.image.wand, '', num)
+        props = [props_p[i] for i in xrange(num.value)]
+        library.MagickRelinquishMemory(props_p)
+        return iter(props)
+
+    def __len__(self):
+        num = ctypes.c_size_t()
+        props_p = library.MagickGetImageProperties(self.image.wand, '', num)
+        library.MagickRelinquishMemory(props_p)
+        return num.value
+
+
 class Sequence(collections.Iterator):
     """MagickWand's image sequences iterator.
 
@@ -1503,4 +1554,3 @@ class ClosedImageError(DestroyedResourceError):
     image.
 
     """
-