+++ docs/changes.rst
@@ -1,6 +1,16 @@
 Wand Changelog
 ==============
 
+Version 0.3.0
+-------------
+
+To be released.
+
+- Added :attr:`Image.metadata <wand.image.Image.metadata>` immutable mapping
+  attribute and :class:`~wand.image.Metadata` mapping type for it.
+  [:issue:`56` by Michael Elovskikh]
+
+
 Version 0.2.2
 -------------